Reorganise to more like my style
