+++ lib/settings/sections/dataset_toggles.dart
@@ -29,8 +29,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:markdown_tooltip/markdown_tooltip.dart';
-import 'package:rattle/settings/widgets/section_partition.dart';
-import 'package:rattle/settings/widgets/section_random_seed.dart';
+import 'package:rattle/settings/sections/partition.dart';
+import 'package:rattle/settings/sections/random_seed.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:rattle/constants/spacing.dart';
