+++ lib/utils/show_settings_dialog.dart
@@ -1,0 +1,52 @@
+/// Show the settings dialog.
+//
+// Time-stamp: <Wednesday 2025-01-15 15:50:09 +1100 Graham Williams>
+//
+/// Copyright (C) 2024, Togaware Pty Ltd
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License");
+///
+/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://www.gnu.org/licenses/>.
+///
+/// Authors: Graham Williams, Kevin Wang
+
+library;
+
+import 'package:flutter/material.dart';
+
+import 'package:rattle/settings/dialog.dart';
+
+void showSettingsDialog(BuildContext context) {
+  showGeneralDialog(
+    context: context,
+    barrierLabel: 'Settings',
+    barrierDismissible: true,
+    barrierColor: Colors.black54, // Darken the background
+    transitionDuration: const Duration(milliseconds: 300),
+    pageBuilder: (context, anim1, anim2) {
+      return const Align(
+        alignment: Alignment.center,
+        child: SettingsDialog(),
+      );
+    },
+    transitionBuilder: (context, anim1, anim2, child) {
+      return FadeTransition(
+        opacity: CurvedAnimation(parent: anim1, curve: Curves.easeOut),
+        child: child,
+      );
+    },
+  );
+}
